Added buttons and function to sorting pets and split the routes into seperate files
