+++ .env
@@ -2,6 +2,6 @@
 DB_NAME='PetAdoptionDB'
 DB_USER='root'
 DB_PASSWORD=''
-Token='eyJ0eXAiOiJKV1QiLCJhbGciOiJSUzI1NiJ9.eyJhdWQiOiI2bGJxb29KeTdDQTlaVUlnY1hWeHNXSERoSTdrcVBpVDZhdTdvVzgyMXNlMXdCd0ZGMSIsImp0aSI6ImI2OGQyZmExNDYxMWMxZWU3NTRkOWZhOTE5ZGI4NGExYTViNDJmMGU3OWE3NTcwMzFmYTc2NjJmMjEwM2JmY2NiMTMxNjU3ZTcyZTRmNGE0IiwiaWF0IjoxNzA3NDUyMjc5LCJuYmYiOjE3MDc0NTIyNzksImV4cCI6MTcwNzQ1NTg3OSwic3ViIjoiIiwic2NvcGVzIjpbXX0.B_8PRYCLTjHrkpK-fnM7MAY7_gQC0tWOcGqDgDjWXfEPz88_KnMs_agKyUQqp5tK_oXrdL72q3po4ybGrQwJz-j8vJWvaYpY0E5Zt7wBJLzKXGmBKYffYCflKNCfmKLwEdOGgUZwU-5xxrrtJnPlilMHA8eMserb41rAN3mw1f7-p3Gi4FF_3n-7j0VjxsCzPmzMGRL4jR00MS2gXhdCeabj8iWAIwhPVsp3WE8TAvaEbVo6Ro3TUTuB6p-OuSVVhIVZk8tCcDw5xXsAQYBu_vU5z2B7xsDYjAJMSQobdJ6I8XE-2qPDc1P_8Y5bw0IS09mqZlZ-7afM5lx40qCq1A'
+PETFINDER_TOKEN='eyJ0eXAiOiJKV1QiLCJhbGciOiJSUzI1NiJ9.eyJhdWQiOiJOUDRySnNTYzNWclZHZ3FFOTE0MEdKVmV5WXlXeGlDRTVGWlE5ZFdhZHR4SGI0M0U4byIsImp0aSI6IjNhODllYTI1ZDc5MzlhODY2YWU0NWFmNjBmNTFjNWVhOWE2OWUyMmVmOTZjNTIyYWFiMzcyMmUyMDQ2ZTFjMDdiYzU3NGEzMjIxZTc4OGRkIiwiaWF0IjoxNzA3NTEyMTM3LCJuYmYiOjE3MDc1MTIxMzcsImV4cCI6MTcwNzUxNTczNywic3ViIjoiIiwic2NvcGVzIjpbXX0.KDk42wc2e32P5nk3h3utIIUxL4vQ1bMiEF1h5ew5LpJfK7ZqaJU-aRQeBnc6RvlF49X0j0EVHg3smlUMxFAMDWv5qBCCB6x94gNn1gRYzy-5GB5YiHDbkXmaO7OMwv0frZCCQWP8NH96xXc8UOK6KLM8Nt6NEBYReTeWJh7aiZ8XH32AUzUA2OROniQvj3alBgVC6pKJd5iqeg1JdSPgLWfSIMpRoMgp6GMXqd2OeqgCEaAUHkOBCD0Lfv7bfRydR1TUfkaYb6Ua59_HabKcrpyEyFFxgUFFqfa7j4ENO8eMjtPzrop_rbIj1igDMftGd0kF17gtR8pvSIgTAcVhOg'
