+++ controllers/api/catRoutes.js
@@ -48,37 +48,6 @@ router.get('/', async (req, res) => {
   }
 });
 
-
-router.get('/pet/:id', async (req, res) => {
-  const petId = req.params.id;
-  try {
-    const petDetail = await axios.get(`https://api.petfinder.com/v2/animals/${petId}`, {
-      headers: {
-        'Authorization': `Bearer ${token}`
-      }
-    });
-
-    const pet = {
-      id: petDetail.data.animal.id,
-      name: petDetail.data.animal.name,
-      gender: petDetail.data.animal.gender,
-      description: petDetail.data.animal.description,
-      breed: petDetail.data.animal.breeds.primary,
-      species: petDetail.data.animal.species,
-      state: petDetail.data.animal.contact.address.state,
-      email: petDetail.data.animal.contact.email,
-      phone: petDetail.data.animal.contact.phone,
-      city: petDetail.data.animal.contact.address.city,
-      image: petDetail.data.animal.photos.length > 0 ? petDetail.data.animal.photos[0].large : null,
-    };
-
-    res.render('pet', pet);
-  } catch (err) {
-    res.status(500).json(err);
-  }
-});
-
-
 router.get('/profile', withAuth, async (req, res) => {
   try {
     const userData = await User.findByPk(req.session.user_id, {
